+++ example/web-app/js/button.element.js
@@ -14,7 +14,7 @@ function ShowListButton() {
   this.clickHandler = function () {
     _state.showList = !_state.showList;
     this.render();
-    UIEvent.dispatch('toggle-list' , "hello world");
+    UIEvent.dispatch('toggle-list');
   }
 
   this.render = function () {